+++ packages/eslint-config-ponder/CHANGELOG.md
@@ -1,4 +1,4 @@
-# eslint-config-ponder
+# eslint-config-ponder-fork
 
 ## 0.0.87
 
@@ -6,4 +6,4 @@
 
 ### Patch Changes
 
-- [#310](https://github.com/0xOlias/ponder/pull/310) [`2ba4b30`](https://github.com/0xOlias/ponder/commit/2ba4b3001402e629dd8956787686a44605c22fa0) Thanks [@eliobricenov](https://github.com/eliobricenov)! - Created `eslint-config-ponder` package and added a prompt to include it during `create-ponder`.
+- [#310](https://github.com/0xOlias/ponder/pull/310) [`2ba4b30`](https://github.com/0xOlias/ponder/commit/2ba4b3001402e629dd8956787686a44605c22fa0) Thanks [@eliobricenov](https://github.com/eliobricenov)! - Created `eslint-config-ponder-fork` package and added a prompt to include it during `create-ponder-fork`.